Refactor project into role-based layers with compatibility facades
