+++ src/cli/__init__.py
@@ -1,0 +1,3 @@
+from src.cli.main import main
+
+__all__ = ["main"]
