+++ src/cli/commands/__init__.py
@@ -1,0 +1,9 @@
+__all__ = [
+    "account",
+    "channel",
+    "collect",
+    "keyword",
+    "scheduler",
+    "search",
+    "serve",
+]
