+++ src/database/__init__.py
@@ -1,0 +1,3 @@
+from src.database.facade import Database
+
+__all__ = ["Database"]
