+++ src/database/connection.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+from pathlib import Path
+
+import aiosqlite
+
+
+class DBConnection:
+    def __init__(self, db_path: str):
+        self._db_path = db_path
+        self.db: aiosqlite.Connection | None = None
+
+    async def connect(self) -> aiosqlite.Connection:
+        Path(self._db_path).parent.mkdir(parents=True, exist_ok=True)
+        self.db = await aiosqlite.connect(self._db_path)
+        self.db.row_factory = aiosqlite.Row
+        return self.db
+
+    async def close(self) -> None:
+        if self.db:
+            await self.db.close()
+
+    async def execute(self, sql: str, params: tuple = ()) -> aiosqlite.Cursor:
+        assert self.db is not None
+        return await self.db.execute(sql, params)
+
+    async def execute_fetchall(self, sql: str, params: tuple = ()) -> list:
+        assert self.db is not None
+        return await self.db.execute_fetchall(sql, params)
