+++ src/database/repositories/settings.py
@@ -1,0 +1,21 @@
+from __future__ import annotations
+
+import aiosqlite
+
+
+class SettingsRepository:
+    def __init__(self, db: aiosqlite.Connection):
+        self._db = db
+
+    async def get_setting(self, key: str) -> str | None:
+        cur = await self._db.execute("SELECT value FROM settings WHERE key = ?", (key,))
+        row = await cur.fetchone()
+        return row["value"] if row else None
+
+    async def set_setting(self, key: str, value: str) -> None:
+        await self._db.execute(
+            "INSERT INTO settings (key, value) VALUES (?, ?) "
+            "ON CONFLICT(key) DO UPDATE SET value = excluded.value",
+            (key, value),
+        )
+        await self._db.commit()
