+++ src/web/routes/scheduler.py
@@ -1,25 +1,27 @@
 from fastapi import APIRouter, Request
 from fastapi.responses import HTMLResponse, RedirectResponse
 
+from src.web import deps
+
 router = APIRouter()
 
 
 @router.post("/tasks/{task_id}/cancel")
 async def cancel_task(request: Request, task_id: int):
-    queue = request.app.state.collection_queue
+    queue = deps.get_queue(request)
     await queue.cancel_task(task_id)
     return RedirectResponse(url="/scheduler?msg=task_cancelled", status_code=303)
 
 
 @router.get("/", response_class=HTMLResponse)
 async def scheduler_page(request: Request):
-    sched = request.app.state.scheduler
-    collector = request.app.state.collector
-    db = request.app.state.db
+    sched = deps.get_scheduler(request)
+    collector = deps.get_collector(request)
+    db = deps.get_db(request)
     msg = request.query_params.get("msg")
     tasks = await db.get_collection_tasks()
     search_log = await db.get_recent_searches()
-    return request.app.state.templates.TemplateResponse(
+    return deps.get_templates(request).TemplateResponse(
         request,
         "scheduler.html",
         {
@@ -42,15 +44,13 @@ async def scheduler_page(request: Request):
 async def start_scheduler(request: Request):
     if getattr(request.app.state, "shutting_down", False):
         return RedirectResponse(url="/scheduler?error=shutting_down", status_code=303)
-    sched = request.app.state.scheduler
-    await sched.start()
+    await deps.scheduler_service(request).start()
     return RedirectResponse(url="/scheduler?msg=scheduler_started", status_code=303)
 
 
 @router.post("/stop")
 async def stop_scheduler(request: Request):
-    sched = request.app.state.scheduler
-    await sched.stop()
+    await deps.scheduler_service(request).stop()
     return RedirectResponse(url="/scheduler?msg=scheduler_stopped", status_code=303)
 
 
@@ -58,11 +58,10 @@ async def stop_scheduler(request: Request):
 async def trigger_collection(request: Request):
     if getattr(request.app.state, "shutting_down", False):
         return RedirectResponse(url="/scheduler?error=shutting_down", status_code=303)
-    sched = request.app.state.scheduler
-    collector = request.app.state.collector
+    collector = deps.get_collector(request)
     if collector.is_running:
         return RedirectResponse(url="/scheduler?msg=already_running", status_code=303)
-    await sched.trigger_background()
+    await deps.scheduler_service(request).trigger_collection()
     return RedirectResponse(url="/scheduler?msg=triggered", status_code=303)
 
 
@@ -70,6 +69,5 @@ async def trigger_collection(request: Request):
 async def trigger_search(request: Request):
     if getattr(request.app.state, "shutting_down", False):
         return RedirectResponse(url="/scheduler?error=shutting_down", status_code=303)
-    sched = request.app.state.scheduler
-    await sched.trigger_search_background()
+    await deps.scheduler_service(request).trigger_search()
     return RedirectResponse(url="/scheduler?msg=search_triggered", status_code=303)
