+++ src/web/routes/search.py
@@ -1,6 +1,8 @@
 from fastapi import APIRouter, Query, Request
 from fastapi.responses import HTMLResponse
 
+from src.web import deps
+
 router = APIRouter()
 
 
@@ -19,42 +21,29 @@ async def search_page(
     offset = (page - 1) * limit
     channel_id_int: int | None = int(channel_id) if channel_id else None
 
-    engine = request.app.state.search_engine
+    service = deps.search_service(request)
 
     if q:
-        if mode == "ai":
-            ai_search = request.app.state.ai_search
-            result = await ai_search.search(q)
-        elif mode == "telegram":
-            result = await engine.search_telegram(q, limit=limit)
-        elif mode == "my_chats":
-            result = await engine.search_my_chats(q, limit=limit)
-        elif mode == "channel":
-            result = await engine.search_in_channel(channel_id_int, q, limit=limit)
-        else:
-            result = await engine.search_local(
-                query=q,
-                channel_id=channel_id_int,
-                date_from=date_from or None,
-                date_to=date_to or None,
-                limit=limit,
-                offset=offset,
-            )
+        result = await service.search(
+            mode=mode,
+            query=q,
+            limit=limit,
+            channel_id=channel_id_int,
+            date_from=date_from or None,
+            date_to=date_to or None,
+            offset=offset,
+        )
 
-    db = request.app.state.db
+    db = deps.get_db(request)
     channels = await db.get_channels()
-    ai_enabled = request.app.state.ai_search.enabled
-
-    # Fetch search quota only when in telegram (Premium) mode
-    search_quota = None
-    if mode == "telegram":
-        search_quota = await engine.check_search_quota()
+    ai_enabled = deps.get_ai_search(request).enabled
+    search_quota = await service.check_quota()
 
     total_pages = 0
     if result and result.total > 0:
         total_pages = (result.total + limit - 1) // limit
 
-    return request.app.state.templates.TemplateResponse(
+    return deps.get_templates(request).TemplateResponse(
         request,
         "search.html",
         {
